Faster (don't search)

+++ calculator.py
@@ -1,3 +1,4 @@
+# Python standard libraries
 from functools import wraps
 import os
 import uuid
@@ -5,28 +6,37 @@ import json
 import sqlite3
 from threading import Thread
 import time
+import timeit
 
+# Flask-y things
 from flask import Flask, request, jsonify, current_app, url_for, render_template
 from flask_jwt_extended import (
     JWTManager, jwt_required, create_access_token,
     get_jwt_identity, verify_jwt_in_request
 )
 from werkzeug.security import safe_str_cmp
-import requests
+from flask_sqlalchemy import SQLAlchemy
 
+# Other libraries
+import requests
 import CoolProp.CoolProp as CP
 import pandas
-
-from flask_sqlalchemy import SQLAlchemy
+import pybtex.database.input.bibtex 
+import pybtex.plugin
+import codecs
+import latexcodec
+from pybtex.style.formatting import plain
 
 app = Flask(__name__)
 
-if os.path.exists('temp.db'):
-    os.remove('temp.db')
+# Flask-SQLAlchemy setup
+# ----------------------
+if os.path.exists('test.db'):
+    os.remove('test.db')
 # app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:' # See this: https://gehrcke.de/2015/05/in-memory-sqlite-database-and-flask-a-threading-trap/
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
-
 class Job(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     uuid = db.Column(db.String)
@@ -43,7 +53,7 @@ jwt = JWTManager(app)
 
 VERIFY = os.environ.get('JWT_VERIFY', False)
 MASTER_KEY = os.urandom(20)
-POLLING_FREQ = 0.1
+POLLING_FREQ = 0.001
 
 def set_verify(value):
     global VERIFY
@@ -52,19 +62,12 @@ def set_verify(value):
 def verification_on():
     return VERIFY
 
-import pybtex.database.input.bibtex 
-import pybtex.plugin
-import codecs
-import latexcodec
-
-from pybtex.style.formatting import plain
 plain_style = plain.Style()
 style = pybtex.plugin.find_plugin('pybtex.style.formatting', 'plain')()
 backend = pybtex.plugin.find_plugin('pybtex.backends', 'html')()
 parser = pybtex.database.input.bibtex.Parser()
 with codecs.open("CoolPropBibTeXLibrary.bib", encoding="latex") as stream:
     data = parser.parse_stream(stream)
-
 # BibTeX_key = 'Span-BOOK-2000'
 # e = list(plain_style.format_entries([data.entries[BibTeX_key]]))[0]
 # print(e.text.render(backend))
@@ -167,35 +170,56 @@ def push_job():
     db.session.commit()
     return jsonify({'uuid': uid})
 
-def _push_job_result(uid, JSONresult):
-    print('pushing result: ', uid)
-    try:
-        job = Job.query.filter_by(uuid=uid).first()
-        db.session.add(Result(uuid=uid, result=json.dumps(JSONresult)))
-        db.session.delete(job)
-        db.session.commit()
-    except sqlite3.IntegrityError as IE:
-        print(IE)
+@app.route('/push_jobs', methods=['POST'])
+@my_jwt_optional
+def push_jobs():
+    uids = []
+    for values in request.get_json():
+        uid = str(uuid.uuid1())
+        db.session.add(Job(uuid=uid, contents=json.dumps(values)))
+        uids.append(uid)
+    db.session.commit()
+    return jsonify({'uuids': uids})
 
 @app.route('/get_results', methods=['POST'])
 @my_jwt_optional
 def get_results():
-    return jsonify({c.uuid:c.result for c in Result.query.all()})
+    tic = timeit.default_timer()
+    while True:
+        # If there are no jobs remaining in the queue,
+        # then return the results, otherwise, see if 
+        # we have timed out yet
+        if not Job.query.all():
+            return jsonify({c.uuid: c.result for c in Result.query.all()})
+        if timeit.default_timer() - tic > 10:
+            raise ValueError()
+        time.sleep(0.01)
 
 def greedy_worker():
     """
-    A single thread that will sequentially run all the jobs that are queued.
-    It does not share with other possible workers
+    A single thread that will sequentially run all the jobs 
+    that are queued. It does not share with other possible workers
     """
     while True:
         time.sleep(POLLING_FREQ)
-        jobs = Job.query.all()
-        for job in jobs:
-            try:
-                j = json.loads(job.contents)
-                _push_job_result(job.uuid, {'output': float(j['Te'])+float(j['Tc'])})
-            except BaseException as BE:
-                print(BE)
+        if Job.query.all():
+            for job in Job.query.all():
+                try:
+                    j = json.loads(job.contents)
+                    JSONresult = {'output': float(j['Te']) + float(j['Tc'])}
+                    db.session.add(Result(uuid=job.uuid, result=json.dumps(JSONresult)))
+                    db.session.delete(job)
+                except BaseException as BE:
+                    print(job.uuid, job.contents, BE)
+            db.session.commit()
+
+@app.route('/flush_results', methods=['POST'])
+@my_jwt_optional
+def flush_results():
+    for r in Result.query.all():
+        db.session.delete(r)
+    db.session.commit()
+    return jsonify({'message':"ok"})
 
 if __name__ == '__main__':
     t = Thread(target=greedy_worker)
